Fix floorplan preview update loop
